adjusted the rotator class to include trigger detection as well as abstracted pickup behaviours
