+++ OPVS-FRIXORIVM/Assets/Prefabs/PickUps/BottleBluePickUp.prefab
@@ -1,5 +1,49 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!1 &412876851222267206
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 3500135899233215793}
+  - component: {fileID: 3208694529532962549}
+  m_Layer: 0
+  m_Name: OnPickUp
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &3500135899233215793
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 412876851222267206}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 292322736742669347}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &3208694529532962549
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 412876851222267206}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a4351acad374dd42b8d99835c39fe22, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &8335498999700606652
 GameObject:
   m_ObjectHideFlags: 0
@@ -13,7 +57,6 @@ GameObject:
   - component: {fileID: 6604429118106032286}
   - component: {fileID: 1851536686600099107}
   - component: {fileID: 1884845498228428035}
-  - component: {fileID: 3666343127565695488}
   m_Layer: 0
   m_Name: BottleBluePickUp
   m_TagString: Untagged
@@ -33,7 +76,8 @@ Transform:
   m_LocalPosition: {x: 0, y: 0.97, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
-  m_Children: []
+  m_Children:
+  - {fileID: 3500135899233215793}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 295.089, y: 90, z: -90}
 --- !u!114 &7873209863190172434
@@ -50,6 +94,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   inclineDegrees: 30
   rotationSpeed: 100
+  actionScript: {fileID: 3208694529532962549}
 --- !u!23 &6604429118106032286
 MeshRenderer:
   m_ObjectHideFlags: 0
@@ -122,16 +167,3 @@ MeshCollider:
   m_Convex: 1
   m_CookingOptions: 30
   m_Mesh: {fileID: 4300000, guid: da8a393ba24fc6f49842d87b0937b84d, type: 3}
---- !u!114 &3666343127565695488
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8335498999700606652}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2a4351acad374dd42b8d99835c39fe22, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  playerTag: Player
