+++ OPVS-FRIXORIVM/Assets/Scripts/GameElements/OnBluePickUp.cs
@@ -4,17 +4,9 @@ using UnityEngine;
 
 public class OnBluePickUp : MonoBehaviour
 {
-    public string playerTag = "Player";
-
-    void OnTriggerEnter(Collider other)
+    public void ItemPickedUp()
     {
-        // Check if the colliding object is the player
-
-        if (other.CompareTag(playerTag))
-        {
-            Destroy(gameObject);
-        }
-
+        Debug.Log("Blue Bottle Picked Up");
     }
 
 }
