+++ OPVS-FRIXORIVM/Assets/Scripts/GameElements/OnScrollPickUp.cs
@@ -6,17 +6,9 @@ using UnityEngine.SceneManagement;
 
 public class OnScrollPickUp : MonoBehaviour
 {
-    public string playerTag = "Player";
-
-    void OnTriggerEnter(Collider other)
+    public void ItemPickedUp()
     {
-        // Check if the colliding object is the player
-
-        if (other.CompareTag(playerTag))
-        {
-            Destroy(gameObject);
-        }
-
+        Debug.Log("Scroll Picked Up");
     }
 
 }
