+++ OPVS-FRIXORIVM/Assets/Scripts/Player/TestController.cs
@@ -34,12 +34,4 @@ public class PlayerController : MonoBehaviour
         rb.AddForce(movement * speed); 
     }
 
-    private void OnTriggerEnter(Collider other)
-    {
-        if(other.gameObject.CompareTag("PickUp"))
-        {
-            other.gameObject.SetActive(false);
-        }
-       
-    }
  }
